+++ src/resources/views/layouts/default/_header_includes.blade.php
@@ -1,0 +1,6 @@
+{{-- Use this partial to add extra scripts, styles for your app --}}
+{{-- The content of this file will be injected in the layout  --}}
+{{-- in the html head section  after AppShell stylesheets --}}
+
+{{-- Create the file resources/views/vendor/appshell/layouts/default/_header_includes.blade.php --}}
+{{-- To populate the content of this view file in your application --}}