+++ tests/Unit/DateTimeWidgetTest.php
@@ -17,7 +17,7 @@ namespace Konekt\AppShell\Tests\Unit;
 use Carbon\Carbon;
 use Konekt\Address\Models\Person;
 use Konekt\AppShell\Tests\TestCase;
-use Konekt\AppShell\Theme\AppShell3Theme;
+use Konekt\AppShell\Theme\AppShellTheme;
 use Konekt\AppShell\Widgets\ShowDate;
 use Konekt\AppShell\Widgets\ShowDateTime;
 use Konekt\AppShell\Widgets\ShowTime;
@@ -28,7 +28,7 @@ class DateTimeWidgetTest extends TestCase
     /** @test */
     public function it_can_render_a_datetime_object()
     {
-        $widget = ShowDateTime::create(new AppShell3Theme());
+        $widget = ShowDateTime::create(new AppShellTheme());
         $output = trim($widget->render(\DateTime::createFromFormat('Y-m-d H:i:s', '2021-04-21 11:27:35')));
 
         $this->assertEquals('2021-04-21 11:27', $output);
@@ -37,7 +37,7 @@ class DateTimeWidgetTest extends TestCase
     /** @test */
     public function it_can_render_a_date()
     {
-        $widget = ShowDate::create(new AppShell3Theme());
+        $widget = ShowDate::create(new AppShellTheme());
         $output = trim($widget->render(\DateTime::createFromFormat('Y-m-d H:i:s', '2021-04-21 11:27:35')));
 
         $this->assertEquals('2021-04-21', $output);
@@ -46,7 +46,7 @@ class DateTimeWidgetTest extends TestCase
     /** @test */
     public function it_can_render_a_time()
     {
-        $widget = ShowTime::create(new AppShell3Theme());
+        $widget = ShowTime::create(new AppShellTheme());
         $output = trim($widget->render(\DateTime::createFromFormat('Y-m-d H:i:s', '2021-04-21 11:27:35')));
 
         $this->assertEquals('11:27', $output);
@@ -55,7 +55,7 @@ class DateTimeWidgetTest extends TestCase
     /** @test */
     public function it_can_render_a_property_of_the_model()
     {
-        $widget = ShowDateTime::create(new AppShell3Theme(), ['text' => '$model.created_at']);
+        $widget = ShowDateTime::create(new AppShellTheme(), ['text' => '$model.created_at']);
         $now = Carbon::now();
         $person = Person::create(['firstname' => 'Johnny', 'lastname' => 'Macaroni']);
         $output = trim($widget->render($person));
@@ -66,7 +66,7 @@ class DateTimeWidgetTest extends TestCase
     /** @test */
     public function the_unknown_date_text_is_a_dash_by_default()
     {
-        $widget = ShowDateTime::create(new AppShell3Theme());
+        $widget = ShowDateTime::create(new AppShellTheme());
         $output = trim($widget->render(null));
 
         $this->assertEquals('-', $output);
@@ -75,7 +75,7 @@ class DateTimeWidgetTest extends TestCase
     /** @test */
     public function the_unknown_date_text_can_be_specified()
     {
-        $widget = ShowDateTime::create(new AppShell3Theme(), ['unknown' => 'never']);
+        $widget = ShowDateTime::create(new AppShellTheme(), ['unknown' => 'never']);
         $output = trim($widget->render(null));
 
         $this->assertEquals('never', $output);