+++ tests/Unit/FilterSetWidgetTest.php
@@ -17,7 +17,7 @@ namespace Konekt\AppShell\Tests\Unit;
 use Konekt\AppShell\Filters\Filters;
 use Konekt\AppShell\Tests\Dummies\SomeFilter;
 use Konekt\AppShell\Tests\TestCase;
-use Konekt\AppShell\Theme\AppShell3Theme;
+use Konekt\AppShell\Theme\AppShellTheme;
 use Konekt\AppShell\Widgets\FilterSet;
 
 class FilterSetWidgetTest extends TestCase
@@ -25,7 +25,7 @@ class FilterSetWidgetTest extends TestCase
     /** @test */
     public function it_can_be_instantiated()
     {
-        $widget = new FilterSet(new AppShell3Theme(), '', new Filters());
+        $widget = new FilterSet(new AppShellTheme(), '', new Filters());
 
         $this->assertInstanceOf(FilterSet::class, $widget);
     }
@@ -33,7 +33,7 @@ class FilterSetWidgetTest extends TestCase
     /** @test */
     public function it_can_be_instantiated_with_the_factory_method_without_any_filters()
     {
-        $widget = FilterSet::create(new AppShell3Theme(), ['route' => '']);
+        $widget = FilterSet::create(new AppShellTheme(), ['route' => '']);
 
         $this->assertInstanceOf(FilterSet::class, $widget);
     }
@@ -41,7 +41,7 @@ class FilterSetWidgetTest extends TestCase
     /** @test */
     public function it_renders_a_form_with_the_route_as_action()
     {
-        $widget = new FilterSet(new AppShell3Theme(), 'appshell.user.index', new Filters());
+        $widget = new FilterSet(new AppShellTheme(), 'appshell.user.index', new Filters());
 
         $html = $widget->render();
         $this->assertStringContainsString('<form', $html);
@@ -52,7 +52,7 @@ class FilterSetWidgetTest extends TestCase
     public function an_array_of_filter_objects_can_be_passed_to_the_constructor()
     {
         $widget = new FilterSet(
-            new AppShell3Theme(),
+            new AppShellTheme(),
             'appshell.user.index',
             new Filters(new SomeFilter('beer_type'))
         );