+++ tests/Unit/TextWidgetTest.php
@@ -17,7 +17,7 @@ namespace Konekt\AppShell\Tests\Unit;
 use Illuminate\Support\Facades\Route;
 use Konekt\AppShell\Models\User;
 use Konekt\AppShell\Tests\TestCase;
-use Konekt\AppShell\Theme\AppShell3Theme;
+use Konekt\AppShell\Theme\AppShellTheme;
 use Konekt\AppShell\Widgets\Text;
 
 class TextWidgetTest extends TestCase
@@ -32,35 +32,35 @@ class TextWidgetTest extends TestCase
     /** @test */
     public function it_can_simply_echo_the_given_text()
     {
-        $text = Text::create(new AppShell3Theme(), ['text' => 'Hello Mirr-murr']);
+        $text = Text::create(new AppShellTheme(), ['text' => 'Hello Mirr-murr']);
         $this->assertEquals('Hello Mirr-murr', trim($text->render()));
     }
 
     /** @test */
     public function it_can_wrap_the_text_into_an_html_tag()
     {
-        $text = Text::create(new AppShell3Theme(), ['text' => 'Yo, Frakk!', 'wrap' => 'span']);
+        $text = Text::create(new AppShellTheme(), ['text' => 'Yo, Frakk!', 'wrap' => 'span']);
         $this->assertEquals('<span>Yo, Frakk!</span>', trim($text->render()));
     }
 
     /** @test */
     public function classes_can_be_added_to_the_wrapper_tag()
     {
-        $text = Text::create(new AppShell3Theme(), ['text' => 'Gemini 6', 'wrap' => 'span', 'class' => 'text-muted']);
+        $text = Text::create(new AppShellTheme(), ['text' => 'Gemini 6', 'wrap' => 'span', 'class' => 'text-muted']);
         $this->assertEquals('<span class="text-muted">Gemini 6</span>', trim($text->render()));
     }
 
     /** @test */
     public function style_can_be_added_to_the_wrapper_tag()
     {
-        $text = Text::create(new AppShell3Theme(), ['text' => 'Gemini 6', 'wrap' => 'span', 'style' => 'color: red']);
+        $text = Text::create(new AppShellTheme(), ['text' => 'Gemini 6', 'wrap' => 'span', 'style' => 'color: red']);
         $this->assertEquals('<span style="color: red">Gemini 6</span>', trim($text->render()));
     }
 
     /** @test */
     public function title_can_be_added_to_the_wrapper_tag()
     {
-        $text = Text::create(new AppShell3Theme(), ['text' => 'Gemini 6', 'wrap' => 'span', 'title' => 'Flew after Gemini 7!']);
+        $text = Text::create(new AppShellTheme(), ['text' => 'Gemini 6', 'wrap' => 'span', 'title' => 'Flew after Gemini 7!']);
         $this->assertEquals('<span title="Flew after Gemini 7!">Gemini 6</span>', trim($text->render()));
     }
 
@@ -68,7 +68,7 @@ class TextWidgetTest extends TestCase
     public function multiple_attributes_can_be_added_to_the_wrapper_tag()
     {
         $text = Text::create(
-            new AppShell3Theme(),
+            new AppShellTheme(),
             [
                 'text' => 'Vostok-3',
                 'wrap' => 'div',
@@ -88,7 +88,7 @@ class TextWidgetTest extends TestCase
     /** @test */
     public function it_can_render_the_given_field_of_a_model_as_text()
     {
-        $text = Text::create(new AppShell3Theme(), ['text' => '$model.name']);
+        $text = Text::create(new AppShellTheme(), ['text' => '$model.name']);
         $user = new User(['name' => 'Mirr Murr']);
         $this->assertEquals('Mirr Murr', trim($text->render($user)));
     }
@@ -96,7 +96,7 @@ class TextWidgetTest extends TestCase
     /** @test */
     public function it_can_render_the_result_of_a_method_of_a_model_as_text()
     {
-        $text = Text::create(new AppShell3Theme(), ['text' => '$model.getEmailForPasswordReset()']);
+        $text = Text::create(new AppShellTheme(), ['text' => '$model.getEmailForPasswordReset()']);
         $user = new User(['email' => 'yo@hey.com']);
 
         $this->assertEquals('yo@hey.com', trim($text->render($user)));
@@ -105,7 +105,7 @@ class TextWidgetTest extends TestCase
     /** @test */
     public function it_can_render_multiple_fields_of_the_model_as_text()
     {
-        $text = Text::create(new AppShell3Theme(), ['text' => '$model.name: $model.getEmailForPasswordReset()']);
+        $text = Text::create(new AppShellTheme(), ['text' => '$model.name: $model.getEmailForPasswordReset()']);
         $user = new User(['name' => 'Johnny', 'email' => 'johnny@macaroni.it']);
 
         $this->assertEquals('Johnny: johnny@macaroni.it', trim($text->render($user)));
@@ -114,7 +114,7 @@ class TextWidgetTest extends TestCase
     /** @test */
     public function it_can_render_fields_containing_underscores_of_the_model_as_text()
     {
-        $text = Text::create(new AppShell3Theme(), ['text' => '$model.is_active']);
+        $text = Text::create(new AppShellTheme(), ['text' => '$model.is_active']);
         $user = new User(['is_active' => 'true']);
 
         $this->assertEquals('true', trim($text->render($user)));
@@ -123,7 +123,7 @@ class TextWidgetTest extends TestCase
     /** @test */
     public function it_can_render_a_combination_of_fields_and_methods_of_the_model_as_text()
     {
-        $text = Text::create(new AppShell3Theme(), ['text' => '$model.name | $model.email']);
+        $text = Text::create(new AppShellTheme(), ['text' => '$model.name | $model.email']);
         $user = new User(['name' => 'Johnny', 'email' => 'johnny@macaroni.it']);
 
         $this->assertEquals('Johnny | johnny@macaroni.it', trim($text->render($user)));
@@ -132,14 +132,14 @@ class TextWidgetTest extends TestCase
     /** @test */
     public function it_can_render_a_route_as_text()
     {
-        $text = Text::create(new AppShell3Theme(), ['text' => ['route' => 'testing-users']]);
+        $text = Text::create(new AppShellTheme(), ['text' => ['route' => 'testing-users']]);
         $this->assertEquals('http://localhost/tesing-users', trim($text->render()));
     }
 
     /** @test */
     public function it_can_render_a_route_with_params_as_text()
     {
-        $text = Text::create(new AppShell3Theme(), ['text' => ['route' => 'testing-user', 'parameters' => ['$model']]]);
+        $text = Text::create(new AppShellTheme(), ['text' => ['route' => 'testing-user', 'parameters' => ['$model']]]);
         $user = new User(['name' => 'Mirr Murr']);
         $user->id = 2733;
         $this->assertEquals('http://localhost/tesing-users/2733', trim($text->render($user)));
@@ -148,24 +148,24 @@ class TextWidgetTest extends TestCase
     /** @test */
     public function a_modifier_can_be_passed_to_manipulate_the_output_of_the_text()
     {
-        $text = Text::create(new AppShell3Theme(), ['text' => 'giovanni gatto', 'modifier' => 'ucwords']);
+        $text = Text::create(new AppShellTheme(), ['text' => 'giovanni gatto', 'modifier' => 'ucwords']);
         $this->assertEquals('Giovanni Gatto', trim($text->render()));
     }
 
     /** @test */
     public function the_modifier_recognizes_registered_widget_modifiers()
     {
-        $text = Text::create(new AppShell3Theme(), ['text' => 'eleonora die hure', 'modifier' => 'uppercase']);
+        $text = Text::create(new AppShellTheme(), ['text' => 'eleonora die hure', 'modifier' => 'uppercase']);
         $this->assertEquals('ELEONORA DIE HURE', trim($text->render()));
 
-        $text = Text::create(new AppShell3Theme(), ['text' => 'Eleonora Die Hure', 'modifier' => 'lowercase']);
+        $text = Text::create(new AppShellTheme(), ['text' => 'Eleonora Die Hure', 'modifier' => 'lowercase']);
         $this->assertEquals('eleonora die hure', $text->render());
 
-        $text = Text::create(new AppShell3Theme(), ['modifier' => 'bool2text:ja,nein']);
+        $text = Text::create(new AppShellTheme(), ['modifier' => 'bool2text:ja,nein']);
         $this->assertEquals('ja', $text->render(true));
         $this->assertEquals('nein', $text->render(false));
 
-        $text = Text::create(new AppShell3Theme(), ['modifier' => 'bool2text:yes,']);
+        $text = Text::create(new AppShellTheme(), ['modifier' => 'bool2text:yes,']);
         $this->assertEquals('yes', $text->render(1));
         $this->assertEquals('', $text->render(null));
     }
@@ -173,7 +173,7 @@ class TextWidgetTest extends TestCase
     /** @test */
     public function a_theme_color_can_be_specified()
     {
-        $text = Text::create(new AppShell3Theme(), ['text' => 'Hey I have color', 'color' => 'primary']);
+        $text = Text::create(new AppShellTheme(), ['text' => 'Hey I have color', 'color' => 'primary']);
 
         $this->assertStringContainsString('class="text-primary', $text->render());
     }
@@ -181,7 +181,7 @@ class TextWidgetTest extends TestCase
     /** @test */
     public function a_concrete_color_can_be_specified()
     {
-        $text = Text::create(new AppShell3Theme(), ['text' => 'Hey I have color', 'color' => '#ff0000']);
+        $text = Text::create(new AppShellTheme(), ['text' => 'Hey I have color', 'color' => '#ff0000']);
 
         $this->assertStringContainsString('style="color: #ff0000', $text->render());
     }